+++ V_Terrain/Code/Utils/Vec3.h
@@ -39,7 +39,7 @@ namespace VTerrain
         bool operator== (Vec3 a) const { return (d[0] == a.d[0] && d[1] == a.d[1] && d[2] == a.d[2]); }
         bool operator!= (Vec3 a) const { return !(*this == a); }
         float LengthSqr() const { return (d[0] * d[0] + d[1] * d[1] + d[2] * d[2]); }
-		float Length() const { return sqrt(d[0]*d[0] + d[1]*d[1] + d[2] * d[2]); }
+		float Length() const { return sqrtf(d[0]*d[0] + d[1]*d[1] + d[2] * d[2]); }
 		void Normalize()
 		{
 			const float len = Length();
@@ -81,8 +81,8 @@ namespace VTerrain
         bool operator!= (Vec2 a) const { return !(*this == a); }
         bool operator< (Vec2 a) const { return (d[0] < a.d[0] ? true : (d[0] == a.d[0] ? (d[1] < a.d[1]) : false)); }
         bool operator> (Vec2 a) const { return (d[0] > a.d[0] ? true : (d[0] == a.d[0] ? (d[1] > a.d[1]) : false)); }
-        float LengthSqr() const { return (d[0] * d[0] + d[1] * d[1]); }
-		float Length() const { return sqrt(d[0]*d[0] + d[1]*d[1]); }
+        float LengthSqr() const { return ((d[0] * d[0] + d[1] * d[1])); }
+		float Length() const { return (sqrtf(d[0]*d[0] + d[1]*d[1])); }
 		void Normalize()
 		{
 			const float len = Length();
